// web/src/lib/api/schemas/route.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Flightpath API
 * Live aircraft positions from the OpenSky Network with flight, airport and airline metadata.
 * OpenAPI spec version: dev
 */
import type { Airport } from "./airport.ts"
import type { MetaSource } from "./metaSource.ts"

export interface Route {
  destination?: Airport
  origin?: Airport
  source: MetaSource
}